Text.Show.Pretty

Description

Functions for human-readable derived Show instances.

Generic representation of values

data Value

Generic Haskell values. NaN and Infinity are represented as constructors. The String in the literals is the text for the literals "as is".

A chain of infix constructors means that they appeared in the input string without parentheses, i.e

1 :+: 2 :*: 3 is represented with InfixCons 1 [(":+:",2),(":*:",3)], whereas

1 :+: (2 :*: 3) is represented with InfixCons 1 [(":+:",InfixCons 2 [(":*:",3)])].

Constructors

Con Name [Value]
Data constructor

InfixCons Value [(Name, Value)]
Infix data constructor chain

Rec Name [(Name, Value)]
Record value

Tuple [Value]
Tuple

List [Value]
List

Neg Value
Negated value

Ratio Value Value
Rational

Integer String
Non-negative integer

Float String
Non-negative floating num.

Char String
Character

String String
String

type Name = String

A name.

valToStr :: Value -> String

Pretty print a generic value. Our intention is that the result is equivalent to the Show instance for the original value, except possibly easier to understand by a human.

valToDoc :: Value -> Doc

Pretty print a generic value. Our intention is that the result is equivalent to the Show instance for the original value, except possibly easier to understand by a human.

valToHtmlPage :: HtmlOpts -> Value -> String

Make an Html page representing the given value.

Values using the Show class

parseValue :: String -> Maybe Value

reify :: Show a => a -> Maybe Value

ppDoc :: Show a => a -> Doc

Try to show a value, prettily. If we do not understand the value, then we just use its standard Show instance.

ppShow :: Show a => a -> String

Convert a generic value into a pretty String, if possible.

Values using the PrettyVal class

dumpDoc :: PrettyVal a => a -> Doc

Render a value in the PrettyVal class to a Doc. The benefit of this function is that PrettyVal instances may be derived automatically using generics.

dumpStr :: PrettyVal a => a -> String

Render a value in the PrettyVal class to a String. The benefit of this function is that PrettyVal instances may be derived automatically using generics.

class PrettyVal a where

A class for types that may be reified into a value. Instances of this class may be derived automatically, for datatypes that support Generics.

Minimal complete definition

Nothing

Methods

prettyVal :: a -> Value

Rendering values to Html

valToHtml :: HtmlOpts -> Value -> Html

Convert a value into an Html fragment.

data HtmlOpts

Options on how to generate Html (more to come).

Constructors

HtmlOpts

Fields

dataDir :: FilePath

Path for extra files. If empty, we look in directory style, relative to document.

wideListWidth :: Int

Max. number of columns in wide lists.

defaultHtmlOpts :: HtmlOpts

Default options.

htmlPage :: HtmlOpts -> Html -> String

Wrap an Html fragment to make an Html page.

newtype Html

Constructors

Html

Fields

exportHtml :: String

Get location of data files

getDataDir :: IO FilePath

Deprecated

ppValue :: Value -> Doc

Deprecated: Please use valToDoc instead.